requirements.txt
numpy==1.24.3
ta
python-binance
pandas
schedule
requests
python-dotenv
